+++ database/migrations/2021_05_01_114906_update_reponse_users_table.php
@@ -15,7 +15,7 @@ class UpdateReponseUsersTable extends Migration
     {
 
         Schema::table('reponse_users', function (Blueprint $table) {
-            $table->unsignedBigInteger('formation_id');
+            $table->unsignedInteger('formation_id');
             $table->foreign('formation_id')->references('id')->on('formations')->onDelete('cascade');
         });
     }
